refactor: centralize document status helper
