+++ app/manage-promoters/[id]/page.tsx
@@ -16,58 +16,16 @@ import {
   BriefcaseIcon,
   ExternalLinkIcon,
   Loader2,
-  ShieldCheckIcon,
-  ShieldAlertIcon,
-  AlertTriangleIcon,
   EditIcon,
 } from "lucide-react"
-import { format, parseISO, differenceInDays, isPast } from "date-fns"
+import { format, parseISO, isPast } from "date-fns"
+import { getDocumentStatus } from "@/lib/document-status"
 import { Separator } from "@/components/ui/separator"
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import LifecycleStatusIndicator from "@/components/lifecycle-status-indicator"
 
 interface PromoterDetails extends Promoter {
   contracts: ContractRecord<{ first_party?: Party; second_party?: Party }>[]
-}
-
-// Re-using helper from manage-promoters page
-const getDocumentStatus = (
-  expiryDate: string | null | undefined,
-): {
-  text: string
-  Icon: React.ElementType
-  colorClass: string
-  tooltip?: string
-} => {
-  if (!expiryDate) {
-    return { text: "No Date", Icon: AlertTriangleIcon, colorClass: "text-slate-500", tooltip: "Expiry date not set" }
-  }
-  const date = parseISO(expiryDate)
-  const today = new Date()
-  const daysUntilExpiry = differenceInDays(date, today)
-
-  if (isPast(date)) {
-    return {
-      text: "Expired",
-      Icon: ShieldAlertIcon,
-      colorClass: "text-red-500",
-      tooltip: `Expired on ${format(date, "MMM d, yyyy")}`,
-    }
-  }
-  if (daysUntilExpiry <= 30) {
-    return {
-      text: "Expires Soon",
-      Icon: ShieldAlertIcon,
-      colorClass: "text-orange-500",
-      tooltip: `Expires in ${daysUntilExpiry} day(s) on ${format(date, "MMM d, yyyy")}`,
-    }
-  }
-  return {
-    text: "Valid",
-    Icon: ShieldCheckIcon,
-    colorClass: "text-green-500",
-    tooltip: `Valid until ${format(date, "MMM d, yyyy")}`,
-  }
 }
 
 function DetailItem({
